requirements.txt
# Prompt Studio - Minimal Dependencies
fastapi==0.116.1
uvicorn==0.35.0
Jinja2==3.1.6
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.0.0